:truck: Rename lexer.ts to parser.ts.
